+++ lana/src/commands/LogView.ts
@@ -2,8 +2,9 @@
  * Copyright (c) 2020 Certinia Inc. All rights reserved.
  */
 import { createReadStream } from 'fs';
+import { writeFile } from 'fs/promises';
 import { basename, dirname, join } from 'path';
-import { WebviewPanel } from 'vscode';
+import { WebviewPanel, window as vscWindow } from 'vscode';
 import { Uri, commands, workspace } from 'vscode';
 
 import { Context } from '../Context';
@@ -15,6 +16,7 @@ interface WebViewLogFileRequest {
   text: string | undefined;
   typeName: string | undefined;
   path: string | undefined;
+  options?: Record<string, never>;
 }
 
 export class LogFileException extends Error {
@@ -70,6 +72,29 @@ export class LogView {
               command: 'getConfig',
               data: workspace.getConfiguration('lana'),
             });
+            break;
+          }
+
+          case 'saveFile': {
+            if (request.text && request.options?.defaultUri) {
+              vscWindow
+                .showSaveDialog({ defaultUri: Uri.file(request.options.defaultUri) })
+                .then((fileInfos) => {
+                  if (fileInfos && request.text) {
+                    writeFile(fileInfos.path, request.text).catch((error) => {
+                      const msg = error instanceof Error ? error.message : String(error);
+                      vscWindow.showErrorMessage(`Unable to save file: ${msg}`);
+                    });
+                  }
+                });
+            }
+            break;
+          }
+
+          case 'showError': {
+            if (request.text) {
+              vscWindow.showErrorMessage(request.text);
+            }
             break;
           }
         }