+++ log-viewer/modules/database-view/DatabaseView.ts
@@ -4,13 +4,25 @@ import { RowComponent, TabulatorFull as Tabulator } from 'tabulator-tables';
 import '../../resources/css/DatabaseView.scss';
 import { DatabaseAccess } from '../Database';
 import '../components/CallStack';
-import { SOQLExecuteBeginLine, SOQLExecuteExplainLine } from '../parsers/TreeParser';
+import Number from '../datagrid/format/Number';
+import { RootNode, SOQLExecuteBeginLine, SOQLExecuteExplainLine } from '../parsers/TreeParser';
 import './DatabaseSOQLDetailPanel';
 import './DatabaseSection';
 
-export function renderDBGrid() {
-  renderDMLTable();
-  renderSOQLTable();
+export async function initDBRender(rootMethod: RootNode) {
+  await DatabaseAccess.create(rootMethod);
+  const dbView = document.getElementById('dbView');
+  if (dbView) {
+    const dbObserver = new IntersectionObserver((entries, observer) => {
+      const visible = entries[0].isIntersecting;
+      if (visible) {
+        observer.disconnect();
+        renderDMLTable();
+        renderSOQLTable();
+      }
+    });
+    dbObserver.observe(dbView);
+  }
 }
 
 function renderDMLTable() {
@@ -46,7 +58,7 @@ function renderDMLTable() {
       dmlData.push({
         dml: dml.text,
         rowCount: dml.rowCount,
-        timeTaken: Math.round((dml.duration / 1000000) * 1000) / 1000,
+        timeTaken: dml.duration,
         timestamp: dml.timestamp,
         _children: [{ timestamp: dml.timestamp, isDetail: true }],
       });
@@ -102,14 +114,30 @@ function renderDMLTable() {
           return 'Total';
         },
       },
-      { title: 'Row Count', field: 'rowCount', sorter: 'number', width: 90, bottomCalc: 'sum' },
+      {
+        title: 'Row Count',
+        field: 'rowCount',
+        sorter: 'number',
+        width: 90,
+        bottomCalc: 'sum',
+        hozAlign: 'right',
+        headerHozAlign: 'right',
+      },
       {
         title: 'Time Taken (ms)',
         field: 'timeTaken',
         sorter: 'number',
         width: 110,
+        hozAlign: 'right',
+        headerHozAlign: 'right',
+        formatter: Number,
+        formatterParams: {
+          thousand: false,
+          precision: 3,
+        },
+        bottomCalcFormatter: Number,
         bottomCalc: 'sum',
-        bottomCalcParams: { precision: 3 },
+        bottomCalcFormatterParams: { precision: 3 },
       },
     ],
     rowFormatter: function (row) {
@@ -140,18 +168,19 @@ function renderDMLTable() {
     }
     table.restoreRedraw();
 
-    const goTo = function () {
-      if (currentSelectedRow) {
-        const nextRow = currentSelectedRow.getNextRow() || currentSelectedRow.getTreeChildren()[0];
-        nextRow &&
+    if (currentSelectedRow) {
+      const nextRow = currentSelectedRow.getNextRow() || currentSelectedRow.getTreeChildren()[0];
+      if (nextRow) {
+        // @ts-expect-error it has 2 params
+        nextRow.scrollTo('center', true).then(() => {
+          //NOTE: This is a workaround for the fact that `row.scrollTo('center'` does not work correctly for ros near the bottom.
+          // This needs fixing in main tabulator lib
           nextRow
             .getElement()
             .scrollIntoView({ behavior: 'auto', block: 'center', inline: 'start' });
+        });
       }
-    };
-    requestAnimationFrame(() => {
-      setTimeout(goTo);
-    });
+    }
   });
 
   // todo: move to a lit element
@@ -214,7 +243,7 @@ function renderSOQLTable() {
         relativeCost: explainLine?.relativeCost,
         soql: soql.text,
         rowCount: soql.rowCount,
-        timeTaken: Math.round((soql.duration / 1000000) * 1000) / 1000,
+        timeTaken: soql.duration,
         aggregations: soql.aggregations,
         timestamp: soql.timestamp,
         _children: [{ timestamp: soql.timestamp, isDetail: true }],
@@ -329,13 +358,14 @@ function renderSOQLTable() {
         width: 120,
         hozAlign: 'right',
         headerHozAlign: 'right',
-        formatter: 'money',
+        formatter: Number,
         formatterParams: {
           thousand: false,
           precision: 3,
         },
+        bottomCalcFormatter: Number,
         bottomCalc: 'sum',
-        bottomCalcParams: { precision: 3 },
+        bottomCalcFormatterParams: { precision: 3 },
       },
       {
         title: 'Aggregations',
@@ -375,18 +405,19 @@ function renderSOQLTable() {
     }
     table.restoreRedraw();
 
-    const goTo = function () {
-      if (currentSelectedRow) {
-        const nextRow = currentSelectedRow.getNextRow() || currentSelectedRow.getTreeChildren()[0];
-        nextRow &&
+    if (currentSelectedRow) {
+      const nextRow = currentSelectedRow.getNextRow() || currentSelectedRow.getTreeChildren()[0];
+      if (nextRow) {
+        // @ts-expect-error it has 2 params
+        nextRow.scrollTo('center', true).then(() => {
+          //NOTE: This is a workaround for the fact that `row.scrollTo('center'` does not work correctly for ros near the bottom.
+          // This needs fixing in main tabulator lib
           nextRow
             .getElement()
             .scrollIntoView({ behavior: 'auto', block: 'center', inline: 'start' });
+        });
       }
-    };
-    requestAnimationFrame(() => {
-      setTimeout(goTo);
-    });
+    }
   });
 
   // todo: move to a lit element